backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langsmith==0.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
ngrok==1.4.0
numpy==1.26.4
openai==1.94.0
orjson==3.10.18
packaging==24.2
pillow==11.3.0
playwright==1.54.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
rag_engine==0.1.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.2.2
tqdm==4.67.1
transformers==4.55.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
